refactor: vo 위치 수정
