+++ mashup-domain/src/main/java/kr/mashup/branding/service/member/dto/MemberCreateVo.java
@@ -1,6 +1,6 @@
 package kr.mashup.branding.service.member.dto;
 
-import kr.mashup.branding.domain.generation.GenerationVo;
+import kr.mashup.branding.service.generation.dto.GenerationVo;
 import kr.mashup.branding.domain.member.Platform;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
